+++ libs/vfs/src/tests/mod.rs
@@ -4,6 +4,7 @@ mod test_cred;
 mod test_dentry;
 mod test_epoll;
 mod test_fdtable;
+mod test_inotify;
 mod test_limits;
 mod test_mount;
 mod test_operation;